another abstract class for tempId to reduce duplicated code
fix tests
