+++ src/main/java/de/tum/in/www1/artemis/domain/quiz/DragItem.java
@@ -1,7 +1,6 @@
 package de.tum.in.www1.artemis.domain.quiz;
 
 import java.util.HashSet;
-import java.util.Objects;
 import java.util.Set;
 
 import javax.persistence.*;
@@ -13,7 +12,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonView;
 
 import de.tum.in.www1.artemis.config.Constants;
-import de.tum.in.www1.artemis.domain.DomainObject;
+import de.tum.in.www1.artemis.domain.TempIdObject;
 import de.tum.in.www1.artemis.domain.view.QuizView;
 import de.tum.in.www1.artemis.service.FilePathService;
 import de.tum.in.www1.artemis.service.FileService;
@@ -24,7 +23,7 @@ import de.tum.in.www1.artemis.service.FileService;
 @Entity
 @Table(name = "drag_item")
 @Cache(usage = CacheConcurrencyStrategy.NONSTRICT_READ_WRITE)
-public class DragItem extends DomainObject {
+public class DragItem extends TempIdObject {
 
     @Transient
     private transient FileService fileService = new FileService();
@@ -52,24 +51,6 @@ public class DragItem extends DomainObject {
     @JsonIgnore
     @Cache(usage = CacheConcurrencyStrategy.NONSTRICT_READ_WRITE)
     private Set<DragAndDropMapping> mappings = new HashSet<>();
-
-    /**
-     * tempID is needed to refer to drag items that have not been persisted yet in the correctMappings of a question (so user can create mappings in the UI before saving new drag
-     * items)
-     */
-    @Transient
-    // variable name must be different from Getter name,
-    // so that Jackson ignores the @Transient annotation,
-    // but Hibernate still respects it
-    private Long tempIDTransient;
-
-    public Long getTempID() {
-        return tempIDTransient;
-    }
-
-    public void setTempID(Long tempID) {
-        this.tempIDTransient = tempID;
-    }
 
     public String getPictureFilePath() {
         return pictureFilePath;
@@ -172,21 +153,6 @@ public class DragItem extends DomainObject {
     public void onDelete() {
         // delete old file if necessary
         fileService.manageFilesForUpdatedFilePath(prevPictureFilePath, null, FilePathService.getDragItemFilepath(), getId());
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        DragItem dragItem = (DragItem) o;
-        if (dragItem.getTempID() != null && getTempID() != null && Objects.equals(getTempID(), dragItem.getTempID())) {
-            return true;
-        }
-        return super.equals(o);
     }
 
     @Override
