+++ src/test/java/de/tum/in/www1/artemis/ExamIntegrationTest.java
@@ -486,8 +486,11 @@ public class ExamIntegrationTest extends AbstractSpringIntegrationBambooBitbucke
         exam.setTitle("Over 9000!");
         long examCountBefore = examRepository.count();
         Exam createdExam = request.putWithResponseBody("/api/courses/" + course1.getId() + "/exams", exam, Exam.class, HttpStatus.CREATED);
-        createdExam.setId(null);
-        assertEquals(exam, createdExam);
+        assertThat(exam.getEndDate()).isEqualTo(createdExam.getEndDate());
+        assertThat(exam.getStartDate()).isEqualTo(createdExam.getStartDate());
+        assertThat(exam.getVisibleDate()).isEqualTo(createdExam.getVisibleDate());
+        // Note: ZonedDateTime has problems with comparison due to time zone differences for values saved in the database and values not saved in the database
+        assertThat(exam).usingRecursiveComparison().ignoringFields("id", "course", "endDate", "startDate", "visibleDate").isEqualTo(createdExam);
         assertThat(examCountBefore + 1).isEqualTo(examRepository.count());
         // No course is set -> conflict
         exam = ModelFactory.generateExam(course1);
