+++ src/test/java/de/tum/in/www1/artemis/RepositoryProgrammingExerciseParticipationResourceIntegrationTest.java
@@ -386,7 +386,9 @@ public class RepositoryProgrammingExerciseParticipationResourceIntegrationTest e
         var receivedLogs = request.getList(studentRepoBaseUrl + participation.getId() + "/buildlogs", HttpStatus.OK, BuildLogEntry.class);
         assertThat(receivedLogs).isNotNull();
         assertThat(receivedLogs).hasSize(1);
-        assertThat(receivedLogs).isEqualTo(logs);
+        assertThat(receivedLogs.get(0).getTime()).isEqualTo(logs.get(0).getTime());
+        // due to timezone assertThat isEqualTo issues, we compare those directly first and ignore them afterwards
+        assertThat(receivedLogs).usingElementComparatorIgnoringFields("time", "id").isEqualTo(logs);
     }
 
     @Test
